move previous results to utils. lint
